fixes issues with person grid
