+++ src/routes/movie/[movieId]/(overview)/index.tsx
@@ -1,7 +1,7 @@
 import { component$ } from "@builder.io/qwik";
 import { DocumentHead } from "@builder.io/qwik-city";
 import { MovieInfoCard } from "~/modules/MovieInfoCard/MovieInfoCard";
-import { PersonCarousel } from "~/modules/PersonGrid/PersonGrid";
+import { PersonCarousel } from "~/modules/PersonCarousel/PersonCarousel";
 import { useMovieLoader } from "../layout";
 
 export default component$(() => {
